完成登陆验证，及部分消息交互

+++ Server/Users.h
@@ -5,7 +5,7 @@
 
 struct AccountData
 {
-	qint16 ID;
+	quint32 ID;
 	QString Name;
 	QString Password;
 };
